# openflare_server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OpenFlare Server 管理端与 Agent API 文档。",
        "title": "OpenFlare Server API",
        "contact": {},
        "version": "3.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/proxy-routes/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProxyRoutes"
                ],
                "summary": "List proxy routes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProxyRoutes"
                ],
                "summary": "Create proxy route",
                "parameters": [
                    {
                        "description": "Proxy route payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.ProxyRouteInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/proxy-routes/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProxyRoutes"
                ],
                "summary": "Update proxy route",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Route ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Proxy route payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.ProxyRouteInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ProxyRoutes"
                ],
                "summary": "Delete proxy route",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Route ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "service.ProxyRouteCustomHeaderInput": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "service.ProxyRouteInput": {
            "type": "object",
            "properties": {
                "cert_id": {
                    "type": "integer"
                },
                "custom_headers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.ProxyRouteCustomHeaderInput"
                    }
                },
                "domain": {
                    "type": "string"
                },
                "enable_https": {
                    "type": "boolean"
                },
                "enabled": {
                    "type": "boolean"
                },
                "origin_url": {
                    "type": "string"
                },
                "redirect_http": {
                    "type": "boolean"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AgentTokenAuth": {
            "description": "Agent API 使用节点专属 Agent Token 或全局 Discovery Token",
            "type": "apiKey",
            "name": "X-Agent-Token",
            "in": "header"
        },
        "BearerAuth": {
            "description": "管理端可使用 Bearer Token，例如：Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}